fix issues with convos backend auth
