+++ .env.example
@@ -9,6 +9,7 @@ PRIVY_APP_CLIENT_ID=your_privy_app_client_id_here
 POSTHOG_API_KEY=your_posthog_api_key_here
 POSTHOG_HOST=your_posthog_host_here
 CONVOS_API_BASE_URL=your_backend_url_here
+FIREBASE_APP_CHECK_TOKEN=firebase_app_check_token
 
 # Add more secrets as needed
 # ANOTHER_SECRET=your_another_secret_here
